oauth: Move views to separate folder
