+++ pkg/webui/oauth/views/app/index.js
