+++ pkg/webui/oauth/views/authorize/authorize.styl
