+++ pkg/webui/oauth/views/create-account/index.js
@@ -26,8 +26,8 @@ import sharedMessages from '../../../lib/shared-messages'
 import Button from '../../../components/button'
 import Field from '../../../components/field'
 import Form from '../../../components/form'
-import IntlHelmet from '../../../lib/components/intl-helmet'
 import Message from '../../../lib/components/message'
+import IntlHelmet from '../../../lib/components/intl-helmet'
 
 import style from './create-account.styl'
 
@@ -107,6 +107,7 @@ export default class CreateAccount extends React.PureComponent {
       this.setState({
         error: '',
         info: getSuccessMessage(result.data.state),
+        registered: true,
       })
     } catch (error) {
       this.setState({
@@ -133,7 +134,9 @@ export default class CreateAccount extends React.PureComponent {
 
     return (
       <div className={style.fullHeightCenter}>
-        <IntlHelmet title={m.register} />
+        <IntlHelmet>
+          <title><Message content={m.register} /></title>
+        </IntlHelmet>
         <div className={style.wrapper}>
           <h1><Message content={m.createAccount} /></h1>
           <Form
